test: update grid visual tests not to use templates
